@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.Concurrent;
+using MyFirstMobileApp.Shared.Models;
+using System.Threading.Tasks;
 
 namespace MyFirstMobileApp.Models
 {
-    public class ItemRepository : IItemRepository
+    public class ItemRepository : IRepository<Item>
     {
         private static ConcurrentDictionary<string, Item> items =
             new ConcurrentDictionary<string, Item>();
@@ -16,41 +18,41 @@ namespace MyFirstMobileApp.Models
             Add(new Item { Id = Guid.NewGuid().ToString(), Text = "Item 3", Description = "This is an item description.", Icon = "https://raw.githubusercontent.com/jamesmontemagno/VS2019-FirstXamarinApp/master/advocado.png" });
         }
 
-        public Item Get(string id)
+        public Task<Item> Get(string id)
         {
-            return items[id];
+            return Task.FromResult(items[id]);
         }
 
-        public IEnumerable<Item> GetAll()
+        public Task<IEnumerable<Item>> GetAll()
         {
-            return items.Values;
+            return Task.FromResult(items.Values as IEnumerable<Item>);
         }
 
-        public void Add(Item item)
+        public Task<bool> Add(Item item)
         {
             item.Id = Guid.NewGuid().ToString();
             items[item.Id] = item;
+            return Task.FromResult(true);
         }
 
         public Item Find(string id)
         {
-            Item item;
-            items.TryGetValue(id, out item);
+            items.TryGetValue(id, out Item item);
 
             return item;
         }
 
-        public Item Remove(string id)
+        public Task<bool> Remove(string id)
         {
-            Item item;
-            items.TryRemove(id, out item);
+            items.TryRemove(id, out Item item);
 
-            return item;
+            return Task.FromResult(true);
         }
 
-        public void Update(Item item)
+        public Task<bool> Update(Item item)
         {
             items[item.Id] = item;
+            return Task.FromResult(true);
         }
     }
 }
