@@ -1,7 +1,4 @@
-﻿using System;
-using Xamarin.Forms;
-using Xamarin.Forms.Xaml;
-using MyFirstMobileApp.Services;
+﻿using Xamarin.Forms;
 using MyFirstMobileApp.Views;
 using Xamarin.Essentials;
 
@@ -9,24 +6,29 @@ namespace MyFirstMobileApp
 {
     public partial class App : Application
     {
+        //"http://myfirstmobileapp-mobileappservice.azurewebsites.net";
         //TODO: Replace with *.azurewebsites.net url after deploying backend to Azure
         //To debug on Android emulators run the web backend against .NET Core not IIS
         //If using other emulators besides stock Google images you may need to adjust the IP address
-        public static string AzureBackendUrl =
+
+        public static string AzureBackendUrl = 
             DeviceInfo.Platform == DevicePlatform.Android ? "http://10.0.2.2:5000" : "http://localhost:5000";
 
-        public static bool UseMockDataStore = false;
+        public static bool UseMockDataStore { get; set; } = false;
 
+
+
+       
         public App()
         {
             InitializeComponent();
 
-            if (UseMockDataStore)
-                DependencyService.Register<MockDataStore>();
-            else
-                DependencyService.Register<AzureDataStore>();
+            Startup.Init();
+
             MainPage = new MainPage();
         }
+
+
 
         protected override void OnStart()
         {
