@@ -3,29 +3,43 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
 
+using Microsoft.Extensions.DependencyInjection;
+
 using Xamarin.Forms;
 
-using MyFirstMobileApp.Models;
+using MyFirstMobileApp.Shared.Models;
 using MyFirstMobileApp.Services;
+using MyFirstMobileApp.Models;
 
 namespace MyFirstMobileApp.ViewModels
 {
     public class BaseViewModel : INotifyPropertyChanged
     {
-        public IDataStore<Item> DataStore => DependencyService.Get<IDataStore<Item>>() ?? new MockDataStore();
+        IRepository<Item> dataStore;
+        public IRepository<Item> DataStore
+        {
+            get
+            {
+                if (dataStore != null)
+                    return dataStore;
+                dataStore = DependencyService.Get<IRepository<Item>>();
+                return dataStore;
+            }
+            set => dataStore = value;
+        }
 
-        bool isBusy = false;
+        bool isBusy;
         public bool IsBusy
         {
-            get { return isBusy; }
-            set { SetProperty(ref isBusy, value); }
+            get => isBusy;
+            set => SetProperty(ref isBusy, value);
         }
 
         string title = string.Empty;
         public string Title
         {
-            get { return title; }
-            set { SetProperty(ref title, value); }
+            get => title;
+            set => SetProperty(ref title, value);
         }
 
         protected bool SetProperty<T>(ref T backingStore, T value,
